+++ examples/hello_xcp_cpp/src/main.cpp
@@ -15,7 +15,7 @@
 // XCP parameters
 
 constexpr const char OPTION_PROJECT_NAME[] = "hello_xcp_cpp";
-constexpr const char OPTION_PROJECT_VERSION[] = __TIME__;
+constexpr const char OPTION_PROJECT_VERSION[] = "V1"; // __TIME__;
 constexpr bool OPTION_USE_TCP = true;
 constexpr uint8_t OPTION_SERVER_ADDR[] = {0, 0, 0, 0};
 constexpr uint16_t OPTION_SERVER_PORT = 5555;
@@ -23,17 +23,42 @@ constexpr size_t OPTION_QUEUE_SIZE = 1024 * 64;
 constexpr int OPTION_LOG_LEVEL = 3;
 
 //-----------------------------------------------------------------------------------------------------
+// Global calibration parameters
+
+// Calibration parameters struct
+struct ParametersT {
+    double min;         // Minimum value for random number generation
+    double max;         // Maximum value for random number generation
+    uint8_t map[4][8];  // Demo 2D map with 4x8 values
+    uint16_t curve[16]; // Demo curve with 16 values
+    uint16_t axis[16];  // Demo axis with 16 values
+};
+
+// Default parameter values
+const ParametersT kParameters = {.min = -2.0,
+                                 .max = +2.0,
+                                 .map = {{0, 1, 2, 3, 4, 5, 6, 7}, {11, 12, 13, 14, 15, 16, 17, 18}, {21, 22, 23, 24, 25, 26, 27, 28}, {31, 32, 33, 34, 35, 36, 37, 38}},
+                                 .curve = {0, 10, 200, 300, 400, 500, 600, 700, 800, 900, 1000, 1100, 1200, 1300, 1400, 1500},
+                                 .axis = {0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15}};
+
+// A global calibration parameter segment handle for struct 'ParametersT'
+// Initialized in main(), after XCP initialization
+std::optional<xcplib::CalSeg<ParametersT>> gCalSeg;
+
+//-----------------------------------------------------------------------------------------------------
 // Demo floating average calculation class
 
 namespace floating_average {
 
-template <uint8_t N> class FloatingAverage {
+class FloatingAverage {
+
+    constexpr static size_t kMaxSamples_ = 64; // Maximum number of samples for the floating average
 
   private:
-    size_t current_index_;          // Current position in the ring buffer
-    uint8_t sample_count_;          // Number of samples collected so far
-    double sum_;                    // Running sum of all samples
-    std::array<double, N> samples_; // Ring buffer for storing samples
+    size_t current_index_;         // Current position in the ring buffer
+    uint8_t sample_count_;         // Number of samples collected so far
+    double sum_;                   // Running sum of all samples
+    double samples_[kMaxSamples_]; // Ring buffer for storing samples
 
   public:
     FloatingAverage();
@@ -41,25 +66,21 @@ template <uint8_t N> class FloatingAverage {
     double calc(double input);
 };
 
-template <uint8_t N> FloatingAverage<N>::FloatingAverage() : samples_{}, current_index_(0), sample_count_(0), sum_(0.0) {
+FloatingAverage::FloatingAverage() : samples_{}, current_index_(0), sample_count_(0), sum_(0.0) {
 
-    // Optional: For measurement of the complete instance, create an A2L typedef for this class (as template, not variant with parameter N)
-    if (A2lOnce()) {
-        A2lTypedefBegin(FloatingAverage, this, "Typedef for FloatingAverage<%u>", N);
-        A2lTypedefMeasurementComponent(current_index_, "Current position in the ring buffer");
-        A2lTypedefMeasurementComponent(sample_count_, "Number of samples collected so far");
-        A2lTypedefMeasurementComponent(sum_, "Running sum of all samples");
-        A2lTypedefEnd();
-    }
-
-    std::cout << "FloatingAverage<" << (unsigned int)N << "> instance created" << std::endl;
+    // Optional: For measurement of the complete instance, create an A2L typedef for this class (once executed and thread safe)
+    A2lCreateTypedef(FloatingAverage, "Typedef for FloatingAverage",                                       //
+                     A2L_MEASUREMENT_COMPONENT(current_index_, "Current position in the ring buffer", ""), //
+                     A2L_MEASUREMENT_COMPONENT(sample_count_, "Number of samples collected so far", ""),   //
+                     A2L_MEASUREMENT_COMPONENT(sum_, "Running sum of all samples", ""),                    //
+                     A2L_MEASUREMENT_ARRAY_COMPONENT(samples_, "Ring buffer for samples", ""));
 }
 
 // Floating average calculate function - instrumented for XCP measurement
-template <uint8_t N> [[nodiscard]] double FloatingAverage<N>::calc(double input) {
+[[nodiscard]] double FloatingAverage::calc(double input) {
 
-    // Calculate the floating average over N samples
-    if (sample_count_ >= N) {
+    // Calculate the floating average over kMaxSamples_ samples
+    if (sample_count_ >= kMaxSamples_) {
         sum_ -= samples_[current_index_];
     } else {
         sample_count_++;
@@ -67,20 +88,17 @@ template <uint8_t N> [[nodiscard]] double FloatingAverage<N>::calc(double input)
     samples_[current_index_] = input;
     sum_ += input;
     double average = sum_ / static_cast<double>(sample_count_);
-    current_index_ = (current_index_ + 1) % N;
+    current_index_ = (current_index_ + 1) % kMaxSamples_;
 
-    // Observe variant N=64
-    // Trigger event 'calc64' and register event and individual parameters, local variables and member variables once
-    if (N == 64) {
-        DaqEventVar(calc64,                                                                     //
-                    A2L_MEAS_PHYS(input, "Input value for floating average", "V", 0.0, 1000.0), // Measure parameter 'input' as physical value in Volt
-                    A2L_MEAS(average, "Current calculated average"),                            // Measure local variable 'average'
-                    A2L_MEAS(current_index_, "Current position in ring buffer"),                // Measure member variable 'current_index_'
-                    A2L_MEAS(sample_count_, "Number of samples collected"),                     // Measure member variable 'sample_count_'
-                    A2L_MEAS(sum_, "Running sum of all samples"),                               // Measure member variable 'sum_'
-                    A2L_MEAS_INST_PTR(FloatingAverage, this, "FloatingAverage",
-                                      "This instance of FloatingAverage")); // Redundant, just to show we also can measure this instance as a whole with a typedef instance
-    }
+    // Trigger event 'calcAvg' and register event and individual parameters, local variables and member variables once
+    DaqEventVar(calcAvg,                                                                    //
+                A2L_MEAS_PHYS(input, "Input value for floating average", "V", 0.0, 1000.0), // Measure parameter 'input' as physical value in Volt
+                A2L_MEAS(average, "Current calculated average"),                            // Measure local variable 'average'
+                A2L_MEAS(current_index_, "Current position in ring buffer"),                // Measure member variable 'current_index_'
+                A2L_MEAS(sample_count_, "Number of samples collected"),                     // Measure member variable 'sample_count_'
+                A2L_MEAS(sum_, "Running sum of all samples"),                               // Measure member variable 'sum_'
+                A2L_MEAS_INST_PTR(FloatingAverage, this, "FloatingAverage",
+                                  "This instance of FloatingAverage")); // Redundant, just to show we also can measure this instance as a whole with a typedef
 
     return average;
 }
@@ -89,19 +107,6 @@ template <uint8_t N> [[nodiscard]] double FloatingAverage<N>::calc(double input)
 
 //-----------------------------------------------------------------------------------------------------
 // Demo random number generator with global calibration parameters
-
-// Calibration parameters for the random number generator
-struct ParametersT {
-    double min; // Minimum random number value
-    double max; // Maximum random number value
-};
-
-// Default parameter values
-const ParametersT kParameters = {.min = -2.0, .max = +2.0};
-
-// A global calibration parameter segment handle for struct 'ParametersT'
-// Initialized in main(), after XCP initialization
-std::optional<xcplib::CalSeg<ParametersT>> gCalSeg;
 
 // Simple random number generation in range [min ..= max] using global calibration parameters
 [[nodiscard]] double random_number() {
@@ -148,7 +153,7 @@ int main() {
 
     // Enable runtime A2L generation for data declaration as code
     // The A2L file will be created when the XCP tool connects and, if it does not already exist on local file system and the version did not change
-    if (!A2lInit(OPTION_SERVER_ADDR, OPTION_SERVER_PORT, OPTION_USE_TCP, A2L_MODE_WRITE_ONCE | A2L_MODE_FINALIZE_ON_CONNECT | A2L_MODE_AUTO_GROUPS)) {
+    if (!A2lInit(OPTION_SERVER_ADDR, OPTION_SERVER_PORT, OPTION_USE_TCP, A2L_MODE_WRITE_ALWAYS | A2L_MODE_FINALIZE_ON_CONNECT | A2L_MODE_AUTO_GROUPS)) {
         std::cerr << "Failed to initialize A2L generator" << std::endl;
         return 1;
     }
@@ -160,54 +165,55 @@ int main() {
     gCalSeg.emplace("Parameters", &kParameters);
 
     // Register the calibration segment description as a typedef and an instance in the A2L file
-    A2lTypedefBegin(ParametersT, &kParameters, "Typedef for ParametersT");
-    A2lTypedefParameterComponent(min, "Minimum random number value", "", -100.0, 100.0);
-    A2lTypedefParameterComponent(max, "Maximum random number value", "", -100.0, 100.0);
-    A2lTypedefEnd();
+    A2lCreateTypedef(ParametersT, "Typedef for ParametersT",                                         //
+                     A2L_MAP_COMPONENT(map, "Demo map", "", 0, 100),                                 //
+                     A2L_CURVE_COMPONENT(curve, "Demo curve", "", 0, 100),                           //
+                     A2L_AXIS_COMPONENT(axis, "Demo axis", "", 0, 100),                              //
+                     A2L_PARAMETER_COMPONENT(min, "Minimum random number value", "", -100.0, 100.0), //
+                     A2L_PARAMETER_COMPONENT(max, "Maximum random number value", "", -100.0, 100.0));
     gCalSeg->CreateA2lTypedefInstance("ParametersT", "Random number generator parameters");
 
     // Create a simple arithmetic local variable
     uint16_t counter{0};
 
     // Create FloatingAverage calculator instances
-    // Example1: On local stack with N=128
-    floating_average::FloatingAverage<128> average_filter1;
-    // Example2: On heap, instance behind a unique_ptr with N=64
-    auto average_filter2 = std::make_unique<floating_average::FloatingAverage<64>>();
-    // Example3:  On heap with N=64
-    auto average_filter3 = new floating_average::FloatingAverage<64>();
+    // Example1: On local stack
+    floating_average::FloatingAverage average_filter1;
+    // Example2: On heap, instance behind a unique_ptr
+    auto average_filter2 = std::make_unique<floating_average::FloatingAverage>();
+    // Example3:  On heap
+    auto average_filter3 = new floating_average::FloatingAverage();
 
     // Main loop
     std::cout << "Starting main loop... (Press Ctrl+C to exit)" << std::endl;
     while (gRun) {
-
         counter++;
         global_counter++;
 
         double input_voltage = random_number();
+        double average_voltage[3];
 
-        // Calculate floating average 128 of input_voltage
-        double average_voltage1 = average_filter1.calc(input_voltage); // Offset input to differentiate from average_filter2
-
-        // Calculate floating average 64 with the 2 instances on heap
-        // Note that the event 'calc64' instrumented inside the FloatingAverage<64>::calc() method, will trigger on each call of any instance (average_filter2 and average_filter3)
+        // Calculate floating averagfe of input_voltage
+        // Note that the event 'calcAvg' instrumented inside the FloatingAverage::calc() method, will trigger on each call of any instance (average_filter2 and average_filter3)
         // Events may be disabled and enabled, to filter out a particular instance to observe
-        DaqEventEnable(calc64);
-        double average_voltage2 = average_filter2->calc(input_voltage - 10.0);
-        DaqEventDisable(calc64);
-        double average_voltage3 = average_filter3->calc(input_voltage + 10.0);
+        DaqEventEnable(calcAvg);
+        average_voltage[0] = average_filter1.calc(input_voltage); // Offset input to differentiate from average_filter2
+        DaqEventDisable(calcAvg);
+
+        // Calculate floating average with the 2 other instances on heap
+        average_voltage[1] = average_filter2->calc(input_voltage - 10.0);
+        average_voltage[2] = average_filter3->calc(input_voltage + 10.0);
 
         // Trigger data acquisition event "mainloop", once register event, global and local variables, and heap instance measurements
-        DaqEventVar(mainloop,                                                                                                                  //
-                    A2L_MEAS(global_counter, "Global counter variable"),                                                                       //
-                    A2L_MEAS(counter, "Local counter variable"),                                                                               //
-                    A2L_MEAS_PHYS(input_voltage, "Input voltage", "V", -100.0, 100.0),                                                         //
-                    A2L_MEAS(average_voltage1, "Calculated average <128> on input_voltage"),                                                   //
-                    A2L_MEAS(average_voltage2, "Calculated average <64>on input_voltage-10"),                                                  //
-                    A2L_MEAS(average_voltage3, "Calculated average <64> on input_voltage+10"),                                                 //
-                    A2L_MEAS_INST(average_filter1, "FloatingAverage", "Local (stack) instance of FloatingAverage<128>"),                       //
-                    A2L_MEAS_INST_PTR(average_filter2, average_filter2.get(), "FloatingAverage", "Heap RAII instance of FloatingAverage<64>"), //
-                    A2L_MEAS_INST_PTR(average_filter3, average_filter3, "FloatingAverage", "Heap instance of FloatingAverage<64>")
+        DaqEventVar(mainloop,                                                                                                 //
+                    A2L_MEAS(global_counter, "Global counter variable"),                                                      //
+                    A2L_MEAS(counter, "Local counter variable"),                                                              //
+                    A2L_MEAS_PHYS(input_voltage, "Input voltage", "V", -100.0, 100.0),                                        //
+                    A2L_MEAS_ARRAY_PHYS(average_voltage, "Calculated average voltages on input_voltage", "V", -100.0, 100.0), //
+                    // Measure all 3 complete instances of FloatingAverage
+                    A2L_MEAS_INST(average_filter1, "FloatingAverage", "Local (stack) instance of FloatingAverage<128>"),                   //
+                    A2L_MEAS_INST_PTR(average_filter2, average_filter2.get(), "FloatingAverage", "Heap RAII instance of FloatingAverage"), //
+                    A2L_MEAS_INST_PTR(average_filter3, average_filter3, "FloatingAverage", "Heap instance of FloatingAverage")
 
         );