+++ inc/a2l.hpp
@@ -13,10 +13,10 @@
 |
  ----------------------------------------------------------------------------*/
 
-// =============================================================================
+// ============================================================================
 // Helper macros for C++ one time and thread safe A2L registrations
 // Mutex protection is needed in multi-threaded contexts, because the A2L registration macros are not thread-safe
-// =============================================================================
+// ============================================================================
 
 /*
 Usage examples:
@@ -102,5 +102,156 @@ template <bool WithMutex, int Location> class A2lOnceGuard<WithMutex, true, Loca
     A2lOnceGuard<true, false, __LINE__> {}
 #define A2lOncePerThread()                                                                                                                                                         \
     A2lOnceGuard<true, true, __LINE__> {}
+
+namespace xcplib {
+
+// =============================================================================
+// Variadic template to create typedefs
+// Once execution and thread safety is handled inside the macro
+// Usage example:
+//   A2lCreateTypedef(ParametersT, "Typedef for ParametersT",
+//                    A2L_PARAMETER_COMPONENT(min, "Minimum value", "miles/h", -0.0, 10.0),
+//                    A2L_PARAMETER_COMPONENT(max, "Maximum value", "miles/h", -70.0, 80.0)
+// );
+// =============================================================================
+
+// Helper lambda-based macros used with A2lCreateTypedef
+
+// Parameters
+#define A2L_PARAMETER_COMPONENT(field_name, comment, unit, min, max)                                                                                                               \
+    [](auto type_ptr) {                                                                                                                                                            \
+        using StructType = std::remove_pointer_t<decltype(type_ptr)>;                                                                                                              \
+        using FieldType = decltype(StructType::field_name);                                                                                                                        \
+        return xcplib::A2lParameterComponentInfo<FieldType>(#field_name, (uint16_t)offsetof(StructType, field_name), 1, 1, comment, unit, min, max);                               \
+    }
+
+// Multi dimensional parameters (curve, map, axis), auto-detect array dimensions (automatic size detection from type)
+#define A2L_CURVE_COMPONENT(field_name, comment, unit, min, max)                                                                                                                   \
+    [](auto type_ptr) {                                                                                                                                                            \
+        using StructType = std::remove_pointer_t<decltype(type_ptr)>;                                                                                                              \
+        using FieldType = decltype(StructType::field_name);                                                                                                                        \
+        using ElementType = std::remove_reference_t<decltype(std::declval<StructType>().field_name[0])>;                                                                           \
+        constexpr size_t x_dim = std::extent_v<FieldType, 0>;                                                                                                                      \
+        return xcplib::A2lParameterComponentInfo<ElementType>(#field_name, (uint16_t)offsetof(StructType, field_name), x_dim, 1, comment, unit, min, max);                         \
+    }
+#define A2L_MAP_COMPONENT(field_name, comment, unit, min, max)                                                                                                                     \
+    [](auto type_ptr) {                                                                                                                                                            \
+        using StructType = std::remove_pointer_t<decltype(type_ptr)>;                                                                                                              \
+        using FieldType = decltype(StructType::field_name);                                                                                                                        \
+        using ElementType = std::remove_reference_t<decltype(std::declval<StructType>().field_name[0][0])>;                                                                        \
+        constexpr size_t x_dim = std::extent_v<FieldType, 0>;                                                                                                                      \
+        constexpr size_t y_dim = std::extent_v<FieldType, 1>;                                                                                                                      \
+        return xcplib::A2lParameterComponentInfo<ElementType>(#field_name, (uint16_t)offsetof(StructType, field_name), x_dim, y_dim, comment, unit, min, max);                     \
+    }
+#define A2L_AXIS_COMPONENT(field_name, comment, unit, min, max)                                                                                                                    \
+    [](auto type_ptr) {                                                                                                                                                            \
+        using StructType = std::remove_pointer_t<decltype(type_ptr)>;                                                                                                              \
+        using FieldType = decltype(StructType::field_name);                                                                                                                        \
+        using ElementType = std::remove_reference_t<decltype(std::declval<StructType>().field_name[0])>;                                                                           \
+        constexpr size_t x_dim = std::extent_v<FieldType, 0>;                                                                                                                      \
+        return xcplib::A2lParameterComponentInfo<ElementType>(#field_name, (uint16_t)offsetof(StructType, field_name), x_dim, 1, comment, unit, min, max);                         \
+    }
+
+// Measurement
+#define A2L_MEASUREMENT_COMPONENT(field_name, comment, unit)                                                                                                                       \
+    [](auto type_ptr) {                                                                                                                                                            \
+        using StructType = std::remove_pointer_t<decltype(type_ptr)>;                                                                                                              \
+        using FieldType = decltype(StructType::field_name);                                                                                                                        \
+        return xcplib::A2lMeasurementComponentInfo<FieldType>(#field_name /*, (FieldType *)nullptr*/, (uint16_t)offsetof(StructType, field_name), 1, comment, unit);               \
+    }
+#define A2L_MEASUREMENT_ARRAY_COMPONENT(field_name, comment, unit)                                                                                                                 \
+    [](auto type_ptr) {                                                                                                                                                            \
+        using StructType = std::remove_pointer_t<decltype(type_ptr)>;                                                                                                              \
+        using FieldType = decltype(StructType::field_name);                                                                                                                        \
+        using ElementType = std::remove_reference_t<decltype(std::declval<StructType>().field_name[0])>;                                                                           \
+        constexpr size_t x_dim = std::extent_v<FieldType, 0>;                                                                                                                      \
+        return xcplib::A2lMeasurementComponentInfo<ElementType>(#field_name /*, (ElementType *)nullptr*/, (uint16_t)offsetof(StructType, field_name), x_dim, comment, unit);       \
+    }
+
+// Typedef
+#define A2L_TYPEDEF_COMPONENT(field_name, type_name, dim)                                                                                                                          \
+    [](auto type_ptr) {                                                                                                                                                            \
+        using StructType = std::remove_pointer_t<decltype(type_ptr)>;                                                                                                              \
+        using FieldType = decltype(StructType::field_name);                                                                                                                        \
+        return xcplib::A2lTypedefComponentInfo<FieldType>(#field_name, type_name, (uint16_t)(dim), (uint16_t)offsetof(StructType, field_name));                                    \
+    }
+
+// Helper struct to hold typedef component information for parameter components
+template <typename FieldType> struct A2lParameterComponentInfo {
+    const char *name;
+    const tA2lTypeId type_id;
+    const uint16_t byte_offset;
+    const uint16_t x_dim;
+    const uint16_t y_dim;
+    const char *comment;
+    const char *unit;
+    double min;
+    double max;
+
+    // Constructor for 2 dimensional array parameter component with physical unit and limit (name, x_dim, y_dim, comment, unit, min, max)
+    constexpr A2lParameterComponentInfo(const char *name, uint16_t byte_offset, uint16_t x_dim, uint16_t y_dim, const char *comment, const char *unit, double min, double max)
+        : name(name), type_id(A2lTypeTraits::GetTypeId<FieldType>()), byte_offset(byte_offset), x_dim(x_dim), y_dim(y_dim), comment(comment), unit(unit), min(min), max(max) {}
+};
+
+// Helper struct to hold typedef component information for measurement components
+template <typename FieldType> struct A2lMeasurementComponentInfo {
+    const char *name;
+    const tA2lTypeId type_id;
+    const uint16_t byte_offset;
+    const uint16_t dim;
+    const char *comment;
+    const char *unit;
+
+    // Constructor for array of physical component (name, dim, comment, unit, min, max)
+    constexpr A2lMeasurementComponentInfo(const char *name, uint16_t byte_offset, uint16_t dim, const char *comment, const char *unit)
+        : name(name), type_id(A2lTypeTraits::GetTypeId<FieldType>()), byte_offset(byte_offset), dim(dim), comment(comment), unit(unit) {}
+};
+
+// Helper struct to hold typedef component information for typedef components
+template <typename FieldType> struct A2lTypedefComponentInfo {
+    const char *name;
+    const char *type_name;
+    const uint16_t dim;
+    const uint16_t byte_offset;
+
+    // Constructor for array of typedef component
+    constexpr A2lTypedefComponentInfo(const char *name, const char *type_name, uint16_t dim, uint16_t byte_offset)
+        : name(name), type_name(type_name), dim(dim), byte_offset(byte_offset) {}
+};
+
+// =============================================================================
+
+// Helper template to register a parameter component
+template <typename T> void A2lCreateTypedefComponentTemplate(const A2lParameterComponentInfo<T> &info) {
+    A2lTypedefParameterComponent_(info.name, info.type_id, info.x_dim, info.y_dim, info.byte_offset, info.comment, info.unit, info.min, info.max, NULL, NULL);
+}
+// Helper template to register a measurement component
+template <typename T> void A2lCreateTypedefComponentTemplate(const A2lMeasurementComponentInfo<T> &info) {
+    A2lTypedefMeasurementComponent_(info.name, info.type_id, info.dim, info.byte_offset, info.comment, info.unit, 0.0, 0.0);
+}
+// Helper template to register a typedef component
+template <typename T> void A2lCreateTypedefComponentTemplate(const A2lTypedefComponentInfo<T> &info) {
+    A2lTypedefComponent_(info.name, info.type_name, info.dim, info.byte_offset);
+}
+
+// Main macro to create a typedef and its fields
+#define A2lCreateTypedef(type_name, comment, ...) xcplib::A2lCreateTypedefTemplate<type_name>(#type_name, sizeof(type_name), comment, __VA_ARGS__);
+
+// Template function for typedef creation
+template <typename TypeName, typename... ComponentBuilders>
+void A2lCreateTypedefTemplate(const char *type_name, size_t type_size, const char *comment, ComponentBuilders &&...builders) {
+    if (XcpIsActivated()) {
+        static std::once_flag once_flag;
+        std::call_once(once_flag, [&]() {
+            A2lLock();
+            A2lTypedefBegin_(type_name, type_size, comment);
+            (A2lCreateTypedefComponentTemplate(builders((TypeName *)nullptr)), ...);
+            A2lTypedefEnd_();
+            A2lUnlock();
+        });
+    }
+}
+
+} // namespace xcplib
 
 #endif // __cplusplus